Move test case code to testing module so it can be mixed in
